backend/requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.2
psycopg[binary]>=3.1.0
pgvector>=0.2.5
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0